Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -14,7 +14,7 @@ class CreateCostumes < ActiveRecord::Migration
   def change
     create_table costumes do |c|
       c.string :name,
-      c.integer :price,
+      c.float :price,
       c.url :url
     end
   end
